Changes naming of primary key constraints to work with single table inheritance.
Currently primary key constraints are based on the name of the org.hibernate.mapping.PrimaryKey object.
However, this leads to duplicated constraint names when using single table inheritance, since the name of these objects seem to match the super-class table name.

Actually, Hibernate itself doesn't even use the name of these PrimaryKey objects
when generating SQL for the creation of primary key constraints (e.g. with hibernate.hbm2ddl.auto=create).
Instead, it seems to leave the naming of the constraint up to the database.
This is why I fully removed the naming approach previously taken from org.hibernate.mapping.PersistentClass, which was probably done with the intention of keeping some kind of consistency
between this plugin and hibernate but which doesn't make sense under the context above.