chore(pom.xml): update liquibase-test-harness and spock-core dependencies
The liquibase-test-harness dependency has been updated from version 1.0.5 to version 1.0.9. This update is necessary to ensure compatibility with the latest changes in the test harness.

The spock-core dependency has been updated from version 2.0-groovy-3.0 to version 2.4-M1-groovy-3.0. This update includes bug fixes and improvements in the Spock testing framework.

These updates will enhance the testing capabilities of the project and ensure compatibility with the latest versions of the dependencies.